time controllers moved to generators
