+++ pygame/controllers/Generator.py
@@ -1,4 +1,21 @@
+import time
 from .Base import ControllerBase
+
+
+class ControllerGeneratorTimeTicks(ControllerBase):
+	def _update(self):
+		return time.perf_counter()
+
+
+class ControllerGeneratorTimeDelta(ControllerBase):
+	def __init__(self, ctrl_time_ticks):
+		super().__init__((ctrl_time_ticks, ))
+		self._last_ticks = 0
+
+	def _update(self):
+		delta = self._in[0].value - self._last_ticks
+		self._last_ticks = self._in[0].value
+		return delta
 
 
 class ControllerGeneratorIncrement(ControllerBase):
