+++ pygame/controllers/__init__.py
@@ -1,10 +1,4 @@
 
 """Controllers package"""
 
-from . import GameLoopState
 from . import *
-from . import Manager
-from . import Input
-from . import Time
-from . import Logic
-from . import Function
